requirements.txt
# Core LangChain & LangGraph
langgraph>=0.2.0
langchain>=0.3.0
langchain-core>=0.3.0
langchain-community>=0.3.0
langchain-openai>=0.2.0
langchain-chroma>=0.1.4
langchain-text-splitters>=0.3.0

# Vector Store
chromadb>=0.5.0

# Document Loaders
pypdf>=4.0.0
arxiv>=2.1.0
wikipedia>=1.4.0

# Web Search
duckduckgo-search>=6.0.0

# API
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
pydantic>=2.9.0
pydantic-settings>=2.6.0

# Observability
langsmith>=0.1.0

# Utils
python-dotenv>=1.0.0
rich>=13.0.0
typer>=0.13.0
pymupdf

# Testing
pytest>=8.0.0
pytest-asyncio>=0.24.0
pytest-cov>=5.0.0
httpx>=0.27.0

# Dev
ruff>=0.7.0
mypy>=1.12.0
